refactor: auth 컴포넌트의 컨테이너 스타일을 공용으로 분리
