+++ src/components/auth/AuthCover.tsx
@@ -1,41 +1,35 @@
 import React from 'react';
 import styled from 'styled-components';
+import { LeftPageContainerStyle } from '../../styles/BookStyleStyles';
 
 export default function AuthCover() {
   return (
-    <AuthCoverStyle>
-      <div className="cover-descriptions">
-        <div className="cover-title">TRIP PLAN</div>
-        <div className="cover-subtitle">여행 투두 리스트를 만들어볼까요?</div>
-      </div>
-      <img src="/src/assets/airplane.png" />
-    </AuthCoverStyle>
+    <LeftPageContainerStyle>
+      <CoverDescriptions>
+        <CoverTitle>TRIP PLAN</CoverTitle>
+        <CoverSubtitle>여행 투두 리스트를 만들어볼까요?</CoverSubtitle>
+      </CoverDescriptions>
+      <CoverImage src="/src/assets/airplane.png" />
+    </LeftPageContainerStyle>
   );
 }
 
-const AuthCoverStyle = styled.div`
-  padding: 0 20px;
-  width: 600px;
-  height: 100%;
-  border-right: 2px solid white;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+const CoverDescriptions = styled.div`
+  text-align: center;
+`;
 
-  .cover-title {
-    font-size: 2rem;
-    font-weight: 700;
-  }
+const CoverTitle = styled.div`
+  font-size: 2rem;
+  font-weight: 700;
+`;
 
-  .cover-subtitle {
-    font-size: 1.2rem;
-    margin-top: 20px;
-  }
+const CoverSubtitle = styled.div`
+  font-size: 1.2rem;
+  margin-top: 20px;
+`;
 
-  img {
-    width: 250px;
-    height: 250px;
-    margin-top: 80px;
-  }
+const CoverImage = styled.img`
+  width: 250px;
+  height: 250px;
+  margin-top: 80px;
 `;
