+++ src/components/auth/LoginForm.tsx
@@ -3,8 +3,16 @@ import { useForm } from 'react-hook-form';
 import { SignupProps } from '../../pages/Signup';
 import { useAuth } from '../../hooks/useAuth';
 import { Link } from 'react-router-dom';
-import { InputStyle, SignUpLink, SignupFormStyle } from './SignupForm';
+import {
+  FormTitle,
+  FormStyle,
+  FieldsetStyle,
+  SubmitButton,
+  InputStyle,
+  SignUpLink,
+} from './StyledComponents';
 import Typography from '../common/Typography';
+import { RightPageContainerStyle } from '../../styles/BookStyleStyles';
 
 export default function LoginForm() {
   const {
@@ -21,10 +29,10 @@ export default function LoginForm() {
   };
 
   return (
-    <SignupFormStyle>
-      <Typography $variant="title1">로그인</Typography>
-      <form onSubmit={handleSubmit(handleSignup)}>
-        <fieldset>
+    <RightPageContainerStyle>
+      <FormTitle $variant="title1">로그인</FormTitle>
+      <FormStyle onSubmit={handleSubmit(handleSignup)}>
+        <FieldsetStyle>
           <InputStyle
             // 개발용 더미 메일
             defaultValue="apple@banana.carrot"
@@ -39,11 +47,11 @@ export default function LoginForm() {
           />
           {isSubmitted && errors.email && (
             <Typography $variant="body1" $color="error">
-              올바른 이메일을 형식을 입력해주세요
+              올바른 이메일을 입력해주세요
             </Typography>
           )}
-        </fieldset>
-        <fieldset>
+        </FieldsetStyle>
+        <FieldsetStyle>
           <InputStyle
             // 개발용 더미 메일
             defaultValue="abc123"
@@ -65,19 +73,19 @@ export default function LoginForm() {
               비밀번호는 6자 이상이어야 합니다
             </Typography>
           )}
-        </fieldset>
-        <button type="submit" disabled={!isValid}>
+        </FieldsetStyle>
+        <SubmitButton type="submit" disabled={!isValid}>
           시작하기
-        </button>
+        </SubmitButton>
         {errorMessage && (
           <Typography $variant="body1" $color="error">
             {errorMessage}
           </Typography>
         )}
-      </form>
+      </FormStyle>
       <SignUpLink>
         계정이 없으신가요? <Link to="/signup">회원가입</Link>
       </SignUpLink>
-    </SignupFormStyle>
+    </RightPageContainerStyle>
   );
 }
