+++ src/pages/Login/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import AuthCover from '../../components/auth/AuthCover';
 import LoginForm from '../../components/auth/LoginForm';
-import BookStyleContainer from '../../components/common/BookStyleContainer';
+import { BookStyleContainer } from '../../styles/BookStyleStyles';
 
 export default function Login() {
   return (
