+++ src/pages/Signup/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import SignupForm from '../../components/auth/SignupForm';
 import AuthCover from '../../components/auth/AuthCover';
-import BookStyleContainer from '../../components/common/BookStyleContainer';
+import { BookStyleContainer } from '../../styles/BookStyleStyles';
 
 export interface SignupProps {
   email: string;
